Adopting new SKIN table in SDE
CCP changed the way skin tables are constructed, namely added another
table skinMaterialSets as a yaml file. The file was posted on #devfleet
on tweetfleet slack by CCP Fox Four, and is supposed to be available in
SDE from now on.

This file is now required to update --skin data using update_yaml.php
script.